--- a/analyzers/tests/SonarAnalyzer.UnitTest/Extensions/ITupleOperationWrapperExtensionsTests.cs
+++ b/analyzers/tests/SonarAnalyzer.UnitTest/Extensions/ITupleOperationWrapperExtensionsTests.cs
@@ -39,8 +39,8 @@ namespace SonarAnalyzer.UnitTest.Extensions
         {
             // NET48 does not support deeply nested tuples and fails with CS8078: An expression is too long or complex to compile
             var deeplyNestedTuple = DeeplyNestedTuple(500); // (1, (2,... , 500))..)
-            // Actual execution time is about 0.5ms.
-            AssertAllElementsExecutionTimeBeLessThan(deeplyNestedTuple, 5.Milliseconds());
+            // Actual execution time is about 0.5 - 10.0 ms.
+            AssertAllElementsExecutionTimeBeLessThan(deeplyNestedTuple, 15.Milliseconds());
 
             static string DeeplyNestedTuple(int depth)
             {