I've recently seen this UT failing randomly on the CI. Last few times with messages:

>Execution of the action should be less than 0.005s, but it required more than 0.009s and 947.9µs.
>Execution of the action should be less than 0.005s, but it required more than 0.012s and 363.9µs.